Puzzle piece turned into a generic script to select puzzle objects Moveable piece inherits from Puzzle Piece
