+++ Assets/Prefabs/Objects/Puzzle-Piece.prefab
@@ -12,8 +12,8 @@ GameObject:
   - component: {fileID: 8760648905007642227}
   - component: {fileID: 3479756754622857480}
   - component: {fileID: 3016530772564845351}
-  - component: {fileID: 310734490987558239}
   - component: {fileID: 4981992808303841672}
+  - component: {fileID: 1979813372965650540}
   m_Layer: 15
   m_Name: Puzzle-Piece
   m_TagString: PuzzlePiece
@@ -108,22 +108,6 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &310734490987558239
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2693792599965057728}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 74e24db7bac8dea43b154e09074444fb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  onSelectYPosition: 2.4
-  isSelected: 0
-  state: 0
-  placedCell: {fileID: 0}
 --- !u!54 &4981992808303841672
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -151,3 +135,20 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 112
   m_CollisionDetection: 0
+--- !u!114 &1979813372965650540
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2693792599965057728}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 47f17d4352a9e0f458b255fc538fc59e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  isSelected: 0
+  state: 0
+  YPostionOffset: 0.25
+  InAirYPosition: 0
+  placedCell: {fileID: 0}
