+++ Assets/Scripts/Puzzles/MoveablePiece.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47f17d4352a9e0f458b255fc538fc59e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
